refactor: changing folder structure.
